Extract the logical realm/instance connection state machine
`session/mod.rs` held the application-side dual-socket flow inline among
scheduler, command and gameplay code. Move that decision layer into a private
`session/connection.rs` and leave transport where it belongs.

Moved, unchanged, as one exact-range relocation (252 lines, 15 items):

  - `poll_instance_link` — attach: the instance channel becomes primary and the
    previous primary is parked as the realm channel.
  - `restore_realm_channels` — restore: logout returns the client to character
    select on the realm connection.
  - `send_packet_realm` / `send_raw_packet_realm` — realm-vs-primary selection.
  - `wait_for_instance_send_before_realm_send_like_cpp` and
    `wait_for_realm_send_before_instance_update_like_cpp`, with the
    `CROSS_SOCKET_WRITE_FENCE_TIMEOUT` they are the only users of.
  - the ConnectTo/endpoint accessors and the two `#[cfg(test)]` installers.

The fence waits move with them because the ordering they preserve is a C++
*session update* invariant: C++ enqueues both `SendDirectMessage` calls inside
one update while Rust drives two independent writer tasks. Deciding that a
realm packet must not overtake an earlier instance packet is application
ordering; the fence that reports physical completion stays network-owned.

`wow-network` keeps sockets, authentication, framing, the encrypted writers,
the byte channels, the FIFO fence primitives and `InstanceLink`. The channel
and fence fields stay on `WorldSession` — retiring that storage is #252/#270,
not this change.

Verified as a pure move: the function multiset across `session/mod.rs` plus the
new module is identical before and after (2626 both sides). The only behavioral
surface change is `poll_instance_link` private -> `pub(super)`, the minimum the
Session driver needs, and it does not widen past the family.

Tests: a routing matrix in `session/tests/connection.rs` covering primary-only,
attach, switch, restore, stale link, pending link, disconnect/failed write and
endpoint round-trip. 7 new tests; the existing routing and instance-link tests
still pass.

Inventories: `poll_instance_link` re-recorded with its new visibility, and the
session hotspot set to its measured rows (production 75031 -> 74866, tests
94834 -> 94946 for the new matrix, total 169865 -> 169812).

Refs #133, #140, #177. Closes #182.
