+++ crates/wow-world/src/session/mod.rs
@@ -7,6 +7,7 @@
 //! [`WorldSocket`](wow_network::WorldSocket) and dispatches them to handlers.
 
 mod admission;
+mod connection;
 pub mod directory;
 mod dispatch;
 pub mod mailbox;
@@ -247,10 +248,6 @@ use wow_social::group::{
     PendingInvites, group_guid_by_db_store_id_like_cpp,
 };
 
-// TrinityCore enqueues cross-connection sends without waiting for physical TCP
-// progress. RustyCore waits briefly to retain the order observed in captures,
-// then completes the already-committed gameplay fanout if a writer stalls.
-const CROSS_SOCKET_WRITE_FENCE_TIMEOUT: Duration = Duration::from_millis(250);
 const QUEST_OBJECTIVE_ITEM_LIKE_CPP: u8 = 1;
 const QUEST_OBJECTIVE_CURRENCY_LIKE_CPP: u8 = 4;
 const QUEST_OBJECTIVE_MIN_REPUTATION_LIKE_CPP: u8 = 6;
@@ -35788,25 +35785,9 @@ impl WorldSession {
         self.session_mgr = Some(mgr);
     }
 
-    /// Set the instance server address and port.
-    pub fn set_instance_endpoint(&mut self, addr: [u8; 4], port: u16) {
-        self.instance_address = addr;
-        self.instance_port = port;
-    }
-
     /// Get the session manager reference.
     pub fn session_mgr(&self) -> Option<&Arc<SessionManager>> {
         self.session_mgr.as_ref()
-    }
-
-    /// Get the instance server address.
-    pub fn instance_address(&self) -> [u8; 4] {
-        self.instance_address
-    }
-
-    /// Get the instance server port.
-    pub fn instance_port(&self) -> u16 {
-        self.instance_port
     }
 
     /// Set the player loading GUID (ConnectTo flow).
@@ -35884,37 +35865,9 @@ impl WorldSession {
         self.player_logout_like_cpp
     }
 
-    /// Set the ConnectTo key.
-    pub fn set_connect_to_key(&mut self, key: Option<i64>) {
-        self.connect_to_key = key;
-    }
-
-    /// Set the ConnectTo serial.
-    pub fn set_connect_to_serial(
-        &mut self,
-        serial: Option<wow_packet::packets::auth::ConnectToSerial>,
-    ) {
-        self.connect_to_serial = serial;
-    }
-
-    /// Set the instance link receiver.
-    pub fn set_instance_link_rx(
-        &mut self,
-        rx: Option<tokio::sync::oneshot::Receiver<InstanceLink>>,
-    ) {
-        self.instance_link_rx = rx;
-    }
-
     /// Get a clone of the send channel.
     pub fn send_tx(&self) -> &flume::Sender<Vec<u8>> {
         &self.send_tx
-    }
-
-    /// Install the FIFO completion fence paired with the session's initial
-    /// realm socket. Runtime does this immediately after constructing the
-    /// session; unit sessions that never own a physical writer leave it empty.
-    pub fn set_send_write_fence_like_cpp(&mut self, fence: SocketWriteFenceLikeCpp) {
-        self.send_write_fence_like_cpp = Some(fence);
     }
 
     pub(crate) fn is_addon_registered_like_cpp(&self, prefix: &str) -> bool {
@@ -38388,62 +38341,6 @@ impl WorldSession {
         }
     }
 
-    /// Poll the instance link oneshot. When received, swap channels and
-    /// continue the player login on the instance socket.
-    async fn poll_instance_link(&mut self) {
-        let rx = match self.instance_link_rx.as_mut() {
-            Some(rx) => rx,
-            None => return,
-        };
-
-        // Non-blocking check
-        match rx.try_recv() {
-            Ok(link) => {
-                info!(
-                    "Instance link received for account {}, swapping channels",
-                    self.account_id
-                );
-
-                // Keep the old realm channels alive — if either TCP connection
-                // drops the WoW client disconnects the whole session.
-                // The realm reader/writer tasks hold the other ends of these
-                // channels, so keeping these receivers/senders prevents the
-                // realm socket from closing.
-                let old_send_tx = std::mem::replace(&mut self.send_tx, link.send_tx);
-                self.realm_send_tx = Some(old_send_tx);
-                let next_write_fence = link
-                    .send_write_fence_like_cpp
-                    .or_else(|| self.send_write_fence_like_cpp.clone());
-                let old_write_fence =
-                    std::mem::replace(&mut self.send_write_fence_like_cpp, next_write_fence);
-                self.realm_send_write_fence_like_cpp = old_write_fence;
-
-                if let Some(pkt_rx) = link.pkt_rx {
-                    let old_packet_rx = std::mem::replace(&mut self.packet_rx, pkt_rx);
-                    self.realm_packet_rx = Some(old_packet_rx);
-                }
-
-                self.instance_link_rx = None;
-
-                // Continue the player login sequence on the instance socket
-                self.handle_continue_player_login().await;
-            }
-            Err(tokio::sync::oneshot::error::TryRecvError::Empty) => {
-                // Not ready yet, keep waiting
-            }
-            Err(tokio::sync::oneshot::error::TryRecvError::Closed) => {
-                warn!(
-                    "Instance link channel closed for account {} — instance connection failed",
-                    self.account_id
-                );
-                self.instance_link_rx = None;
-                self.player_loading = None;
-                self.connect_to_key = None;
-                self.release_character_login_claim_like_cpp();
-            }
-        }
-    }
-
     /// Check for area triggers at the player's current position.
     ///
     /// This is called after movement updates to handle:
@@ -39417,120 +39314,6 @@ impl WorldSession {
         self.send_packet(&packet);
     }
 
-    /// Send a server packet on the **realm** connection.
-    ///
-    /// Some packets (e.g. `QueryPlayerNamesResponse`) must travel on the
-    /// realm socket, not the instance socket.  Falls back to `send_tx` if
-    /// no realm channel exists (pre-ConnectTo or single-connection mode).
-    pub fn send_packet_realm(&self, pkt: &impl wow_packet::ServerPacket) {
-        let data = pkt.to_bytes();
-        let tx = self.realm_send_tx.as_ref().unwrap_or(&self.send_tx);
-        if tx.send(data).is_err() {
-            warn!("Realm send channel closed for account {}", self.account_id);
-        }
-    }
-
-    /// Wait for current-instance packets to reach their physical socket before
-    /// emitting a later realm packet. C++ enqueues both `SendDirectMessage`
-    /// calls during one session update; Rust's two independent writer tasks
-    /// need this FIFO completion fence to retain observed cross-connection
-    /// order.
-    pub(crate) async fn wait_for_instance_send_before_realm_send_like_cpp(&self) -> bool {
-        let Some(realm_send_tx) = self.realm_send_tx.as_ref() else {
-            return true;
-        };
-        if realm_send_tx.same_channel(&self.send_tx) {
-            return true;
-        }
-        let Some(write_fence) = self.send_write_fence_like_cpp.as_ref() else {
-            warn!(
-                account = self.account_id,
-                "instance/realm ordering fence unavailable for separate sockets"
-            );
-            return false;
-        };
-        match write_fence
-            .wait_for_prior_packets_written_like_cpp(
-                &self.send_tx,
-                CROSS_SOCKET_WRITE_FENCE_TIMEOUT,
-            )
-            .await
-        {
-            SocketWriteFenceWaitResultLikeCpp::Written => true,
-            SocketWriteFenceWaitResultLikeCpp::TimedOut => {
-                warn!(
-                    account = self.account_id,
-                    timeout_ms = CROSS_SOCKET_WRITE_FENCE_TIMEOUT.as_millis(),
-                    "instance writer did not acknowledge the ordering fence before timeout"
-                );
-                false
-            }
-            SocketWriteFenceWaitResultLikeCpp::WriterClosed => {
-                warn!(
-                    account = self.account_id,
-                    "instance writer closed before acknowledging the ordering fence"
-                );
-                false
-            }
-        }
-    }
-
-    /// Wait for realm packets to reach their physical socket before emitting a
-    /// later instance update. This mirrors `Player::SendNewItem` preceding the
-    /// deferred `Map::SendObjectUpdates` player-field flush in C++.
-    pub(crate) async fn wait_for_realm_send_before_instance_update_like_cpp(&self) -> bool {
-        let Some(realm_send_tx) = self.realm_send_tx.as_ref() else {
-            return true;
-        };
-        if realm_send_tx.same_channel(&self.send_tx) {
-            return true;
-        }
-        let Some(write_fence) = self.realm_send_write_fence_like_cpp.as_ref() else {
-            warn!(
-                account = self.account_id,
-                "realm/instance ordering fence unavailable for separate sockets"
-            );
-            return false;
-        };
-        match write_fence
-            .wait_for_prior_packets_written_like_cpp(
-                realm_send_tx,
-                CROSS_SOCKET_WRITE_FENCE_TIMEOUT,
-            )
-            .await
-        {
-            SocketWriteFenceWaitResultLikeCpp::Written => true,
-            SocketWriteFenceWaitResultLikeCpp::TimedOut => {
-                warn!(
-                    account = self.account_id,
-                    timeout_ms = CROSS_SOCKET_WRITE_FENCE_TIMEOUT.as_millis(),
-                    "realm writer did not acknowledge the ordering fence before timeout"
-                );
-                false
-            }
-            SocketWriteFenceWaitResultLikeCpp::WriterClosed => {
-                warn!(
-                    account = self.account_id,
-                    "realm writer closed before acknowledging the ordering fence"
-                );
-                false
-            }
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn install_realm_send_channel_for_test(&mut self, tx: flume::Sender<Vec<u8>>) {
-        self.realm_send_tx = Some(tx);
-    }
-
-    #[cfg(test)]
-    pub(crate) fn install_realm_send_write_fence_for_test(
-        &mut self,
-        fence: SocketWriteFenceLikeCpp,
-    ) {
-        self.realm_send_write_fence_like_cpp = Some(fence);
-    }
-
     /// Send pre-serialized packet bytes to the client.
     ///
     /// Used for packets with dynamic opcodes (e.g. `SetSpellModifier`
@@ -39550,19 +39333,6 @@ impl WorldSession {
         }
         if self.send_tx.send(data.to_vec()).is_err() {
             warn!("Send channel closed for account {}", self.account_id);
-        }
-    }
-
-    /// Send pre-serialized packet bytes on the realm connection.
-    ///
-    /// This is the cross-session counterpart of [`Self::send_packet_realm`]:
-    /// registry commands already carry serialized bytes, but C++ opcode
-    /// routing still requires packets such as `SMSG_PARTY_INVITE` and
-    /// `SMSG_PARTY_MEMBER_FULL_STATE` to use `CONNECTION_TYPE_REALM`.
-    pub(crate) fn send_raw_packet_realm(&self, data: &[u8]) {
-        let tx = self.realm_send_tx.as_ref().unwrap_or(&self.send_tx);
-        if tx.send(data.to_vec()).is_err() {
-            warn!("Realm send channel closed for account {}", self.account_id);
         }
     }
 
@@ -57390,39 +57160,6 @@ impl WorldSession {
     /// Whether the session is disconnecting.
     pub fn is_disconnecting(&self) -> bool {
         self.state == SessionState::Disconnecting
-    }
-
-    /// Restore the realm socket as the primary send/receive channel.
-    ///
-    /// After a ConnectTo flow, `send_tx` and `packet_rx` point to the
-    /// instance socket while the realm channels are stored in
-    /// `realm_send_tx` / `realm_packet_rx`.  On logout the client
-    /// returns to character select on the REALM connection, so we must
-    /// swap back.  The old instance channels are simply dropped — the
-    /// instance reader/writer tasks will notice and exit.
-    pub(crate) fn restore_realm_channels(&mut self) {
-        if let Some(realm_tx) = self.realm_send_tx.take() {
-            info!(
-                "Restoring realm send channel as primary for account {}",
-                self.account_id
-            );
-            self.send_tx = realm_tx;
-        }
-        if let Some(realm_write_fence) = self.realm_send_write_fence_like_cpp.take() {
-            self.send_write_fence_like_cpp = Some(realm_write_fence);
-        }
-        if let Some(realm_rx) = self.realm_packet_rx.take() {
-            info!(
-                "Restoring realm packet channel as primary for account {}",
-                self.account_id
-            );
-            self.packet_rx = realm_rx;
-        }
-        // Clear any pending ConnectTo state
-        self.instance_link_rx = None;
-        self.connect_to_key = None;
-        self.connect_to_serial = None;
-        self.player_loading = None;
     }
 }
