+++ docs/architecture/ownership-and-boundaries.md
@@ -118,6 +118,7 @@ last-writer-wins policy.
 | Concept | Current owner and storage | Writers | Readers / delivery | Clock, lifetime, and synchronization | Retirement |
 |---|---|---|---|---|---|
 | Authenticated world connection | `wow-network::accept` and the connection task | socket/authentication task | `WorldSession` dispatch boundary | one connection; created after authentication and dropped on disconnect | Remains a network responsibility. #134 narrowed the listener to transport-owned configuration and authenticated connection outputs. |
+| Logical realm/instance connection | private `wow_world::session::connection`: the attach/switch/restore state machine, realm-vs-primary send selection and the two cross-socket ordering fences | the owning Session task, which is the only writer of the logical primary | `send_packet_realm`/`send_raw_packet_realm` callers, the Session driver polling the instance link, and logout restore | connection lifetime; after `SMSG_CONNECT_TO` the instance channel becomes primary and the realm channel is parked, never closed — the client drops the session if either socket dies | #182 extracted the transition implementation out of `session/mod.rs`. The physical channels and fences stay on `WorldSession` fields for now; #252 and #270 retire that storage. |
 | Gameplay session construction aggregate | private `SessionResources` in `crates/world-server/src/session_resources.rs` | `world-server` bootstrap constructs it; the outer session callback captures and clones the `Arc` | `world-server::create_session` copies stores, registries, DB adapters, and runtime handles into each `WorldSession`; the aggregate never enters `wow-network` | process lifetime; no independent clock | #134 moved the aggregate to composition code; #136 moved the process into a library-backed composition root and extracted the private factory. Capability-specific dependency reduction remains in the physical and ownership lanes. |
 | Session directory | opaque `wow_world::session::directory::PlayerRegistry`; its `DashMap<ObjectGuid, compatibility entry>` storage is private to the owner module | named generation-aware registration/unregistration owns lifecycle; temporary state-publication compatibility methods remain counted | #192 runtime/fanout, #193 combat/loot and #194 quest/spell/movement consumers use owned bounded results and generation-checked delivery; remaining social/group access is assigned exactly in [`player-registry-consumer-map.tsv`](player-registry-consumer-map.tsv) | process lifetime; presence, incarnation and addressability are Session concerns, while gameplay fields are temporary mirrors | #150 installed the opaque lifecycle seam; #192-#194 closed their consumer slices; #195 burns down remaining production consumers; #196 removed compatibility storage APIs and fixtures; #138 relocated the already-opaque directory to the transitional `wow-world` owner, leaving the terminal `wow-session::directory` move open. |
 | Session mailbox and durable rails | private `wow_world::session::mailbox::{protocol, durable, pump}`: `SessionCommand` plus every ordinary payload, `SharedClientVisibleGuidsLikeCpp`, the durable creature-runtime rail and the single pump. #189 moved the durable loot-money coordinator to `wow_world::loot_persistence` | session/global runtime and gameplay producers enqueue commands; one session task consumes them | the owning session consumes FIFO commands and publishes acknowledgements | connection/session lifetime; queue identity, capacity, FIFO, incarnation fences, acknowledgements, and shutdown drain are observable | #191 and #190 completed their bounded protocol/rail relocations, #138 closed broad directory access, and #140 moved the complete vertical without changing queue identity, FIFO, durability, capacity, incarnation fences, acknowledgements or shutdown drain. `wow-network` now owns only transport primitives. #189 still removes durable loot persistence coordination. |
