+++ tools/architecture/runtime-ownership-ledger.json
@@ -90,7 +90,7 @@
           "session_key",
           "session_mgr"
         ],
-        "current_storage": "Physical socket channels/fences and ConnectTo state are stored directly on WorldSession; socket tasks and SessionManager remain in wow-network.",
+        "current_storage": "Physical socket channels/fences and ConnectTo state are stored directly on WorldSession; socket tasks and SessionManager remain in wow-network. Since #182 the logical attach/switch/restore transitions and realm-vs-primary routing over those fields live in the private wow_world::session::connection module rather than session/mod.rs.",
         "sole_writer": "The connection/session task owns logical routing state; wow-network connection tasks own physical channel and fence lifecycle.",
         "readers": "WorldSession send/routing helpers, ConnectTo/login flow, shutdown, and network writers.",
         "clock_lifetime": "One authenticated connection with realm and optional instance links; physical links live until disconnect and both sockets must remain alive after ConnectTo.",
@@ -1869,9 +1869,9 @@
       "entries": [
         {
           "path": "crates/wow-world/src/session/mod.rs",
-          "production_lines": 75031,
-          "test_lines": 94834,
-          "total_lines": 169865,
+          "production_lines": 74866,
+          "test_lines": 94946,
+          "total_lines": 169812,
           "owner": "wow-world WorldSession/application monolith.",
           "open_retirement_issues": [
             140,
